+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG for Sulu
 * dev-develop
     * ENHANCEMENT #3312 [SecurityBundle]        Added configuration options to send reset password message 
     * ENHANCEMENT #3317 [All]                   Fixed jackalope conflict in composer.json
+    * ENHANCEMENT #3315 [RouteBundle]           Added route-history controller
     * ENHANCEMENT #3306 [DocumentManagerBundle] Adapted document-manager service definition for new NodeNameSlugifier
     * ENHANCEMENT #3302 [ContentBundle]         Added metadata to configure remove-live
     * FEATURE     #3300 [ContentBundle]         Added onInvalid flag with ignore option to properties